parse xml only if xml parseable in invoiceimporter (affecting metrics)
